ofx package rename, added support for custom/promo splashscreen, added natron api pdf, readded compat libs, move manifest to logs and other minor changes. SDK2 for workshop is more or less complete, will start testing the binaries, and continue on my todo list.

+++ installer/scripts/build-natron.sh
@@ -103,6 +103,10 @@ else
   cat $CWD/installer/freebsd/config.pri > config.pri || exit 1
 fi
 
+if [ "$CUSTOM_SPLASH" == "1" ] && [ -f $CWD/installer/misc/splashscreen.png ]; then
+  cp $CWD/installer/misc/splashscreen.png Gui/Resources/Images/ || exit 1
+fi
+
 rm -rf build
 mkdir build || exit 1
 cd build || exit 1
